recipes-core/images: Update RESIN_BOOT_PARTITION_FILES
Remove imx-boot-${MACHINE}-sd.bin-flash_evk from boot
partition files for the Balena image and add boot.scr
to the Balena flasher image.
The flash_evk is not used on the boot partition of the
flasher image. The boot.scr is the script that re-writes
the bootloader using the USB as source

Changelog-entry: Update RESIN_BOOT_PARTITION_FILES
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>