linux-balena-bootloader: append required command line args
On devices like the iot-gate-imx8 where the bootROM is not able to boot
from an external device, the bootROM boots into the internal bootloader
and then reads from the external device.

Hence, we can't expect that the bootloader will always be able to pass
the balena bootloader's arguments and we extend the balena bootloader's
kernel with them instead so that the balena bootloader is always
booted to behave correctly.

Change-type: patch
Changelog-entry: extend balena bootloader's kernel command line arguments
Signed-off-by: Alex Gonzalez <alexg@balena.io>