[REVIEW] initscript: abroot: rollback u-boot on rootfs switch
The current linux-compulab kernel freezes when booting from kexec as
the ocotp clocks are accessed before being enabled - this is a problem
when trying to rollback from altboot failed HUP to a balena bootloader enabled
image as the old kernel will freeze on the kexec.

This commit adds a rollback function that on partition switch reinstates
the old OS U-boot which bypasses the balena bootloader.

Signed-off-by: Alex Gonzalez <alexg@balena.io>