imx-boot: do not use remove on tasks
The :remove bitbake keyword is meant to do word removals and here it is
used to remove a whole line, which works, but could break at any time.

Replace by a prepend that defines the missing variables to fix the
original install task.

Changelog-entry: fix imx-boot not to use remove on the install task
Signed-off-by: Alex Gonzalez <alexg@balena.io>