os-helpers: add os-helpers-power to customize shutdown
The flasher script in meta-balena calls `poweroff` and expects it to
fail if the device cannot be powered off. However on this device type
the call returns success but the device is not powered off.

This commit uses a custom shutdown that effectively shuts down this
device type.

Changelog-entry: customize shutdown
Signed-off-by: Alex Gonzalez <alexg@balena.io>