workflows: iot-gate-imx8-sb: deploy to specific device type
The iot-gate-imx8-sb workflow builds an iot-gate-imx8 with secure boot
support, the machine is still iot-gate-imx8, but it's meant to deploy
to its own iot-gate-imx8-sb device type as the original iot-gate-imx8
device type rootfs does not fit the secure boot support and there is
no migration path at the moment.

However, it's desired to keep this a single machine as once there is a
migration path the same device type will support both secure boot and
non secure boot enabled variants.

Changelog-entry: let the iot-gate-imx8-sb workflow deploy to a specified slug
Signed-off-by: Alex Gonzalez <alexg@balena.io>