hostapp-update-hooks: guarantee emmc boots from boot0
Make sure the emmc partition config is set to look for the bootloader
in the boot0 partition. This is set on Compulab's boot script when
updating u-boot but we stopped using it for balenaOS in
f202fbd33335ca7fde7ad695fadea24ced6f77d1.

Changelog-entry: make sure emmc boots from partition boot0
Signed-off-by: Alex Gonzalez <alexg@balena.io>