linux-balena-bootloader_5.15.32.bb: Fixes for usrmerge feature
When enabling the usrmerge feature we will have /usr/lib populated.
So until then, we need to prepare for it by changing the modules removal
code to use a variable for doing that and also explicitly remove /usr also.

Changelog-entry: Switch to using nonarch_base_libdir in the linux-balena-bootloader recipe
Signed-off-by: Florin Sarbu <florin@balena.io>